+++ src/forge_io/readers/__init__.py
@@ -2,5 +2,6 @@
 
 from forge_io.readers.arri_reader import ArriRawReader
 from forge_io.readers.oiio_reader import OIIOReader
+from forge_io.readers.red_reader import RedRawReader
 
-__all__ = ["ArriRawReader", "OIIOReader"]
+__all__ = ["ArriRawReader", "OIIOReader", "RedRawReader"]